Trim processed_parameters list.
It was initially created to circumvent problems that arose because
config items and data items were under the shared postfix pillar and
not under the postfix:config as well as the postfix:aliases pillar.

The list was then further expanded to include settings necessary for
the broken virtualuser feature which broke generic mapping support.

Trimming the list down to nothing for preparation of a truly generic
mapping support.